+++ src/js/06-tab-animations-and-folder-bounce.js
@@ -128,6 +128,7 @@
     try {
       gURLBar.setURI();
     } catch (err) {
+      noteError("tab animations and folder bounce: hideWwwInUrlbar", err);
     }
   }